make the apply branch configurable
The workflow was originally written to run a `terraform plan` on
`pull_request` actions and `terraform apply` on a `push` to the
`master` branch.

Not all of our TF repos use `master` as their default branch
though. Newer ones are going to be `main` and Wharf still mostly
operates on `develop`.

This adds an `apply_branch` input to the workflow to allow this to be
set. It defaults to `master` so it will be backwards compatible with
existing repos that use it.