add terraform drift detection workflow
There are a few TF projects that I'd like to have automated drift
detection running on. With this workflow, we can add a github action
to run it daily and then github actions will send us an alert any time
`terraform plan` shows any changes to be made (ie, our TF code is out
of sync with what's running).